Work: new class
The Work holds data necessary for management of tasks executed
in a different context then the main application flow. It is
a persistable component so it can be optionally stored in
database.

Each Work instance holds only data and no actual processing
code. The code is to be implemented in a WorkExecutor.

Accessing internal data of Work must be synchronized with RWLock.
Usually, some GUI would need to ask the state of a Work and
the WorkExecutor would update the Work. Use accessLock() for this
purpose. If multiple fields of a Work are to be accessed, it is
possible to hold the accessLock() at once and pass SKIP_LOCK
to the access methods.

The Work::id() is not a subject of locking as it MUST NEVER EVER
be changed during runtime, once it has been set. Only Dao layer
usually access this field.

A single Work instance must not be executed in multiple threads
concurrently. To avoid such case, the executionLock() is provided.
If a Work is to be executed, the lock must be held first.

Each Work is in one of the following states:

* STATE_IDLE - never executed, never scheduled, created
* STATE_SCHEDULED - scheduled in a queue
* STATE_EXECUTED - being executed, can be (re)scheduled
	(but executionLock() must be held)
* STATE_FINISHED - finished, cannot be scheduled nor executed
* STATE_FAILED - failed, cannot be scheduled nor executed
* STATE_CANCELED - work has been canceled prior its finish

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>