WorkExecutor: new class
The Work holds information necessary for a task processing.
The actual Work execution routine is implemented by a WorkExecutor.
There is usually a pair (Work, WorkExecutor) which work together.

The WorkExecutor::accepts() tells whether the given WorkExecutor
instance understands the given Work instance. The WorkExecutor
is usually a singleton with no persisting capabilities. It can
however initiate database transactions or access other resources
in the system.

The WorkExecutor is always executed with Work::executionLock() held.
It MUST lock the accessLock() when accessing the Work contents.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>