WorkSuspendThrowable: new class
The only purpose of this class is to simplify suspending flow of
a work in the middle of its execution code path. It is not derived
from any exception-like object to avoid anybody to catch it
accidentaly instead of the actual suspend operation.

The WorkSuspendForEventThrowable is an alternative way of the
suspend operation where the work is suspended to wait for an
external event only.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>