About: describe program purpose and requirements
An About instance allows to describe requirements on the Poco library
and some details about the current program being started.

The Poco library version can be checked automatically for requirements
to assure we are running against the right version. This is might help
to mitigate security flaws.

The program can define its own version if she knows it. Also a description
can be supplied this way.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>