DIDaemon: new class
The DIDaemon is a common implementation of the Poco::Util::ServerApplication
intended for most programs based on the DependencyInjector. Such program
creates a single instance of LoopRunner and starts it.

DIDaemon implements several standard command line options that should cover
most requirements of all applications. Currently there are the following
options:

* -h  print help and exit
* -V  print version and exit
* -D  override a configuration option
* -c  load a configuration file
* -N  send SIGTERM to a program after the daemon has started

The LoopRunner instance is search via DependencyInjector. The name of instance
being started can be specified via a configuration file or -D option.

The DIDaemon should be started via its static method up() as:

  DIDaemon::up(argc, argv);

to handle unexpected exceptions and signals. Additional information can be
passed vi the About struct.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>