DependencyInjector: fix delete order
DependencyInjector should delete instances in the reverse order
then in which they have been created. This behaviour has been
broken when migrating to the DIWrapper. In fact, it is possible
the order has been broken even before because the instances
are inserted into the m_free vector in post-order way. The
current instance being created is inserted last.

Fixes: 626c08239504 ("DependencyInjector: manage instances wrapped in SharedPtr")