AbstractWorkRepository: new abstract class
Most real WorkRepository implementations would need to solve
the persistance mechanism of Work inheritance tree. As there
can be a great count of Work implementations for different
tasks, the repository must be able to handle this fact.

The AbstractWorkRepository utilizes WorkFactory to determine
type of each Work and to create the right Work instance when
loading them.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>