Backtrace: new class
Backtrace allows to collect the function call backtrace independently
on the compiler, libc and platform. If there is no backtrace support
available, it provides no information and Backtrace::support is zero.

Currently, only glibc library is supported. However, the backtrace(3)
function might be avaiable in other libraries as well (e.g. uclibc).
As there is currently no way how to test it, no other libc library is
currently considered.

On ARM, the backtrace may behave in a different way so the tests must
be fixed for this architecture.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Martin Sakin <xsakin00@stud.fit.vutbr.cz>