BasicQueue: new class
Basic synchronous queue of Work. It orders Work instances based on
the Work::activationTime() and Work::priority(). If the activation
time is NULL, such Work instance is considered to be waiting for
an event and thus is not activated until something wakes it up.

The BasicQueue provides unlocked interface for manipulation with
scheduling. The pushUnlocked, wakeupUnlocked and cancelUnlocked
can schedule, wakeup or cancel a Work in the context of the queue.
The BasicQueue::lock() must always be held by the caller.

The BasicQueue::pop() interface is either blocking or non-blocking
and the locking is done internally. It returns the first Work
that is to be executed and removes it from the queue.

Several parts of the queue locks Work::accessLock() to modify the
Work instance and Work::executionLock() to avoid interfering with
the execution stage (i.e. reschedule while executing).

The BasicQueue consists of 2 collections:

* m_active - multimap of {time, priority} -> Work::Ptr
* m_queue - map of WorkID -> {Work::Ptr, ActiveQueue::iterator}

The m_active collection contains all Work instances having the
activation time specified and so they are expected to execute.
It must be multimap as there might be two works with exactly
the same execution time and priority. The order of work issuing
is implemented inside BasicQueue::Key and respects the activation
time and priority. The implementation should avoid starving of
low-priority works.

The m_queue collection contains all scheduled works (also, the
ones waiting for an event, thus inactive ones). It also contains
the iterator into m_active that informs whether the work is
active and allows to quickly deactivate it if needed.

The blocking behaviour depends on the state of the queue. If there
is some active work, it blocks until the activation time of the
first active work is reached. Otherwise, it blocks indefinitely
(empty queue). Anybody can call BasicQueue::notify() to stop
blocking and reconsider the situation. If the queue is woken up
by notify(), it does not block anymore and thus it allows to
quit the application gratefully.

Every Work instance that entered the BasicQueue ends up in state
Work::STATE_SCHEDULED. Work instances entering the queue with
states Work::STATE_FINISHED, Work::STATE_FAILED or Work::STATE_CANCELED
are ignored.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>