WorkRepository: new interface
We need a way how to persist and load back Work instances to cover
persistent works. If a work should survive an application restart,
it has to be stored (usually) in a database. The WorkRepository
provides an interface to do this.

An EmptyWorkRepository implementation is provided for cases where
works are always non-persistent.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>