WorkRunner: new abstract class
Provide an abstract class for implementing the Work execution flow.
The WorkRunner is a kind of an interface that allows a WorkScheduler
implementation to execute works in threads. A WorkRunner is always
created via its WorkRunnerFactory and thus different WorkRunner
implementations can be used.

Each WorkRunner implementation must ensure locking the WorkExecution
guard prior the actual execution of a Work.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>