GenericWorkRunner: new class
Generic implementation of a WorkRunner. It implements the work execution
flow. While executing a work, its executionLock() lock is held. The
GenericWorkRunner handles suspend requests and uncaught Exceptions.

It marks the executed works as Work::STATE_EXECUTED while executing
and then Work::STATE_FINISHED or Work::STATE_FAILED based on the
wait the work exists. If a work asks for suspend, then it calls
the associated WorkScheduler to reschedule the work.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>