JsonUtil: fix handling of null for double
The JSON specification (see json.org) does not define format for
INF or Not-a-Number values. Thus, when parsing a double, the null
value might be used to represent those. Define that with JsonUtil,
the null means NaN. Provide a template specialization for double
to handle the null value and return NaN as expected.

Fixes: 3b43b165e44e ("JsonUtil: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>