DevicePrefix: new class
A DevicePrefix is a single byte value that is used to determine a device
type with respect to the technology it uses for communication. The
prefix is used to decide which Gateway module should handle it.

There is a list of well-known and (potentially) supported technologies.
Thus, the DevicePrefix is represented as an Enum.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>