CryptoConfig: new class
When cryptography is needed in an application, we have to manage
its configuration. Usually, we encrypt and decrypt many different
data messages that has its specific parameters (CryptoParams).
However, we need to inject the secret passphrase to generate an
encryption key.

This class provides methods for this purpose. Any application class
(usually some service or frontend class) can use the CryptoConfig
class to provide the always required configuration entries. It is
expected that the properties:

* algorithm - name of the cipher
* passphrase - secret passphrase
* iterationCount - number of iterations performed by the cipher

are set once e.g. via dependency injection and the application
object would receive them for using.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Reviewed-by: Ondřej Hujňák <ihujnak@fit.vutbr.cz>