WorkAccess: introduce Work-specific locking mechanism
Each Work instance can be (from the locking point of view) in one
of two states:

* IDLE - it exists, it is somewhere (probably scheduled in some
         execution queue) but not executing.

* EXEC - it was scheduled and now it is executing.

The two states are important to consider when locking the work
efficiently. Each work is always concurrently readable. If there
is some writer, reading is not possible. This is ensured by the
common RWLock (readers-writers lock).

Accessing the Work is thus possible only by using some guard class
derived from WorkAccess: WorkReading or WorkWriting. Both ensures
proper locking of the work for single or block atomic operations.
There is however one aspect to consider, the used RWLock is not
recursive a thus it must not be locked more then once in a single
thread.

When a work is in the EXEC state, no writer except of the work's
execution thread can modify the work internals. During execution,
the work must not be changed from the outer world otherwise the
results might be undefined. Especially, it is not possible to cancel
an executing work. The work leaves the execution as EXECUTED, FAILED
or FINISHED. If somebody would modify the state during its execution,
this state change will be probably discarded when leaving the execution.

To avoid such situations, there is another lock called execution lock.
This lock is recursive and it is held during the work's execution via
the WorkExecution class. A particular WorkRunner ensures to hold this
lock before a work is executed. The WorkWriting guard first tries to
lock the execution lock and then it races for the writing lock.

On the contrary, the WorkReading does not lock the execution lock
because reading of work is always possible unless there is an active
writer of the RWLock.

Thus, using of WorkWriting and WorkReading is mandatory (and enforced
by the Work interface) to assure that locking works as expected.
Also, specialized Work classes (derived from Work) should use those
locks to make sure that access of their content does not introduce
race conditions.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>