StoppableLoopAdapter: fix memory leak
The test for running thread is not a good idea here as it
prevents to join the thread and then delete it. The method
doStop() is called from both stop() and from the destructor
and thus it may left the m_thread allocated.

The original intention was probably to avoid reallocation
of the thread on every stop. However, this is stupid and
not very useful.

Valgrind results:
HEAP SUMMARY:
    in use at exit: 92 bytes in 2 blocks
  total heap usage: 258,712 allocs, 258,710 frees, 24,853,071 bytes allocated

Fixes: e175707c4d6a ("Import of code extracted from the server project")
Fixes: a337c90280ec ("StoppableLoop: new class")

Reported-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>