di: support setter accepting argument char
Introduce the class DIWCharSetter to support setter methods accepting
a char argument. It is derived from DIWTextSetter so the macro
BEEEON_OBJECT_TEXT is suitable for this purpose and no other changes
are necessary.

In case when the text passed to a DIWCharSetter is longer then a single
character, the DIWWrongInputException is thrown.

The DIWrapperTest is updated to test this feature accordingly.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>