Console: introduce a new concept
The Console holds some common properties of a console and
provides mechanisms for session creation. When a session
is created, it can be used for communication with the user.

Handlers of a console input work with the ConsoleSession.
Console implementations can support multiple parallel sessions
or multiple sequentially allocated sessions.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>