TCPConsole: new Console implementation
Implements Console API for TCP communication. Note, that the TCPConsole
does not implement any security layer. Thus, it is not suitable for a
serious network communication. The purpose of this class is to enable
a debugging or testing support for applications that cannot communicate
via stdin and stdout directly (background processes).

The TCPConsole listens on 127.0.0.1:6000 as default.

Note that the TCPConsoleTest is very limited due to a non-deterministic
behaviour of socket connections in some environment.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>