Printable: extract new interface
Both Console and SimpleConsole are in fact a kind of an output
stream. Extract interface Printable that defines the method
print and allows to write a polymorphic code that does not
depend on the type of console used.

The IOSPrintable allows to wrap std::ostream to be Printable
and thus improves the polymorphic potential.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>