ArgsParser: new class
The class implements parsing of arguments and respects quotation marks.
It is intended as a replacement of the Poco::StringTokenizer that can
only handle a simple separator (a space).

The ArgsParser handles arguments enclosed in quotes (configurable symbol)
"argument". Such argument can contain whitespaces (configurable symbol set)
"argument and text" and it also supports escaping of quotes like
"text \"in quotes\"". Escaping of escape symbol is also possible "\\".

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>