Enum: fix random generation for sparse enums
In case when the enum is defined as a sequence of sparse values, e.g.:

enum {
	X0 = 5,
	X1 = 10,
	X2 = 11,
	X3 = 200,
	...
}

the random() method fails because it assumes that the values are always
starting at 0 and are contiguously incremented by 1.

Fix this behaviour by iterating over the Base::valueMap() to find the
n'th enum value based on a random number. The random generation thus
has a linear complexity. However, most enums are not so big that this
performance issue would bother us.

A unit test of the random() method is included to show that this
implementation of the method random() is fixed.

Fixes: ac930f3c84b2 ("DeviceID: the prefix rewitten in enum")
Reported-by: Hošala Martin <xhosal00@stud.fit.vutbr.cz>
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>