DeviceID: avoid generation of random prefix
The DeviceID does not generate any random prefix. This way, it is
possible to generate a random device ID with any arbitrary prefix.
The caller may generate the prefix at random as well.

This fixes also a potential inconsistence that could happen when
passing PREFIX_INVALID and expect a random valid prefix to be
generated. We could end up again with PREFIX_INVALID.

Fixes: ac930f3c84b2 ("DeviceID: the prefix rewitten in enum")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>