ConfigurationSaver: new class
The ConfigurationSaver is a counter-part to the ConfigurationLoader.
Based on the configured file name, it determines the right
implementation of the AbstractConfiguration and provides it to the
caller. The caller fills the configuration with data and then asks
the ConfigurationSaver to write it to the file. It hides the file-type
detection from the application.

Includes also static method to save the given configuration in the
given file.

Supported types of configutration: XML, JSON, PropertyFile.

Signed-off-by: Martin Hošala <xhosal00@stud.fit.vutbr.cz>
[used ClassInfo::forPointer]
Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>