DIWrapperFactory: new class
As a future replacement of the ManifestSingleton, the DIWrapperFactory
is introduced. It is an interface for creation of DIWrapper instances.
It also manages a global list of available factories. Factories would
be similarly to the method used for ManifestSingleton. However, we
avoid calling the Poco macros anymore. This simplifies the code and
improves readability a bit.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>