WorkLockManager: new class
The WorkLockManager is a central point for managing locks for Works.
The goal is to get rid of locks placed inside the Work instances.
Locks are managed via the HashedLock allocator. This might possibly
introduce deadlocks in case that somebody attempts to lock more
Work instances at once.

The execution locks are to be used by the WorkRunner only. It ensures
that a single Work instance cannot be executed mutliple times in
different threads.

The class would replace the WorkAccess-based locking.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>