Work: remove WorkAccess arguments in setter/getter methods
This code was enforcing holding of locks for accessing the Work content.
However, it is quite inpractical for using and complicates other development.
Due to the included locks, it is difficult to provide some WorkDao or
other mechanisms that do not use Work::Ptr but maintain just a working
copies.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>