DeviceInfoProvider: method findByNameAndVendor
The main purpose of this method is to provide a way how to match devices
discovered by the gateway with those in the devices.xml or other devices
storage without using just indices (some magic numbers deviced by the
former magicians ).

The gateway reports e.g.:

  new-device "Some cool company" "Very nice sensor"

and we can lookup the registered devices for it.

Signed-off-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>