GWMessage: new class
The GWMessage is abstract class representing messages (including
their contents), that are being sent between Gateway and Server.
Messages are transfered in the JSON format.

The internal representation of a GWMessage is the Poco::JSON::Object.
All subclasses should set the values directly into that instance via
accessing method json(). This leads to a lazy-validation behaviour.

Signed-off-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>