GWListenRequest: new class
The message is sent by the server to the gateway and it is intended
to start "listening mode" on the gateway. In this mode, gateway can
discover new devices from the sensoric network.

The listening mode has typically a limited duration. After the
duration exceeds, the listening mode is automatically turned off.

Signed-off-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>