DependencyInjector: call hook cleanup in destructor
Provide a way how to notify instances about their deinitialization.
When the DependencyInjector is being destroyed, all registered
instances are about to be deleted to. This may or may not happen
depending on the state of the shared pointer of each instance.
If somebody still holds the pointer, then such instance is not
destroyed (its destructor is not called). However, the hook
cleanup is called always.

The testCircularDependency() shows that the cleanup hook can be
used to solve circular dependencies. If two objects holds a shared
pointer to each other, then their destructors are never called.
The cleanup hook can be used to break the dependency when the
DependencyInjector is being destroyed and thus clean up properly.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>