FdStream: new class
Wrap file descriptor into std::istream and std::ostream interfaces. Thus,
it is now possible to work with file descriptors in the C++ way. Also
non-blocking approach is possible, a non-blocking read sets eof() when
there is nothing to read.

The FdStream can be used in select/poll/epoll approaches and it directly
supports the poll call without any other complex coding. The poll error
events are converted to a FileException.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>