test: utilize TestTimingListener for all tests
The TestTimingListener is always used but it can be configured to
never output anything (this is the default). It outputs the timing
measurements into a file specified by the environment variable
TEST_OUTPUT_TIMING. It recognizes 3 special values:

 * empty string - no output at all
 * "stdout" - measurements are printed to stdout
 * "stderr" - measurements are printed to stderr

Other strings are treated as file names and thus timing report is
printed to a file.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>