DependencyInjector: improve destruction of instances
Instead of just a simple iteration over the vector of allocated instances,
do now a more complex destruction routine. We iterate over the instances
many time and check their reference count to determine whether it is possible
to destroy it or not.

Doing this many times allows to remove a layer of dependency in every attempt.
Finally, only instances with circular dependencies remain. We left those objects
to solve their issues on their own.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>