ClosedConsoleSessionImpl: new class
Define a hidden implementation of ClosedConsoleSessionImpl. Its purpose
is to be always closed. I.e. it returns eof() == true and returns empty
inputs and does no output.

The Console provides an instance of the ClosedConsoleSessionImpl
by calling to Console::closedSession(). This is intended for derived
classes to call when detecting a close request.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Tested-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>