PosixSignal: implement send() to a thread
The implementation is now pthread-only.

In Thread_POSIX.cpp, the Thread tid is created as:

309 #if POCO_OS == POCO_OS_LINUX
310         // On Linux the TID is acquired from the running thread using syscall
311         _pData->tid = 0;
312 #elif POCO_OS == POCO_OS_MAC_OS_X
313         _pData->tid = static_cast<TIDImpl>( pthread_mach_thread_np(_pData->thread) );
314 #else
315         _pData->tid = _pData->thread;
316 #endif
...
461 #if POCO_OS == POCO_OS_LINUX
462         pThreadImpl->_pData->tid = static_cast<TIDImpl>( syscall (SYS_gettid) );
463 #endif

However, in my setup, the Poco::Thread::currentTid() returns a different value
then what the Linux syscall gettid gives. I found out in runtime that the currentTid()
denotes a pthread_t which conforms to be the #else part of the first condition.

However, in application, the POCO_OS == POCO_OS_LINUX is evaluated to
true and thus it is inconsistent.

This code is a subject of future changes to ensure correct function under other
environment settings.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Tested-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>