EventSource: replace exception throw with warning log
If AsyncExecutor was not set every "execute()" call resulted
in throwing the IllegalStateException.

This is not what higher level functions would expect or need
to sanitate, so this behaviour is now replaced by a warning log.
Utilizing Once helper this case is logged only the first time
someone call "execute()".

No operation without AsyncExecutor set is executed either way.

Signed-off-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
[minor coding style changes]
Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>