Translator: remake the method format()
This is a backwards incompatible change. The method format() is
changed to accept also an argument def that is used when the key
is not found for translation. In this case, the formatFallback()
is used to format the def by the given arguments.

The original method format() is renamed to formatSure(). It works
the same way, however, it is intended for core translations that
are for sure always available. Thus, the current code should just
use formatSure() instead of format() and everything will be OK.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>