DIWrapper: check cast is possible during build
When extracting an instance from the given DIWrapper, we assume
that if the wrapper casts to AbstractDIWrapper<T> then the instance
of class T can also be casted to its parent class B.

If T cannot be casted to B, the build should fail here. It would
otherwise throw a bad_cast exception during runtime if executed.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>