DIWrapper: handle possibly failing dynamic_cast
The dynamic_cast of the found setter to the setter expected by the property
name in the XML file ("text", "number", "ref", ...) can throw std::bad_cast
exception. This exception is never thrown and leads to fail of the application
with no useful description or other context. Thus, it is difficult to debug it.

Change the dynamic_cast to cast to a pointer instead of the reference and
handle the failed cast by throwing a well defined exception DIWWrongInputException
with a sensible message.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>