DependencyInjector: utilize DIWCast in create<T>() method
The create<T>() method automatically casts the wanted instance
to the target type instead of getting DIWrapper * pointer and
trying to do it another way.

However, the create<T>() would fail when requesting an instance
of class X : public T. Because, it casts to AbstractDIWrapper<T>
but the current DIWrapper is AbstractDIWrapper<X>. This leads to
sd::bad_cast exception.

By utilizing DIWCast, we can safely cast the wanted instance to
its parent. It works the same way as during the ref injection.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>