JsonUtil: remove method extract()
The method is not a better replacement for Poco::JSON::Object::getValue().
There is however one case when it does matter - processing of doubles.
The JSON format cannot encode NAN or INF values and thus the only way
how to send them is as null. Calling getValue<double>(key) on a null
value leads to Poco::IllegalAccessException. This can be handled by
calling optValue<double>(key, std::nan("")) instead.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>