MultiException: new class
There are cases when processing a number of values or events in a loop
where each particular iteration of the loop can fail, however, we want
all the loop to finish without interruption.

If any such iteration fails - throws an exception there is a question
what to do. We may log that exception, omit it (not a good idea) or
collect it and after the loop is done, process all the collected
exceptions and make some summarization to the caller. The MultiException
is intended exactly for this purpose.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>