Backtrace: new method dump()
For situations like signal handlers, it might be desirable to
output the backtrace without side-effects. Thus, calling common
logging facilities might not be available. The method dump()
guarantees that dumping of the backtrace into the given file
descriptor is safe for such purposes.

Note that the method backtrace() might call dlopen (malloc) and
thus it is necessary to call it at least once before any crash
to avoid this situation.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>