Zip::Iterator: new class
The class Zip and Zip::Iterator provides an elegant way to
iterate over 2 containers, which size can be different.

The Zip supports both const and non-const containers.

The Zip::Iterator returns std::pair of references to values
of containers and provides methods for returning iterators
of containers. It also provides special functions to find out
which of the iterators did not come to an end.

Signed-off-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>
Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>