InfoProvider: parseFile is not virtual anymore
There seems to be no reason to define a virtual method parseFile
inside InfoProvider as this polymorphism was never used. When we
call parseFile (from XmlInfoProvider), we already know the file
format and there is no need to be polymorphic. The method was
protected so the outer world could not use it.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>