DevicePrefix: normalize enum names
All prefix names are now lower-case in a programming-style. The enum
values are not intended to be in human-style (for UX). For UX, we should
have translation layers that performs sufficient conversions if needed.

By using the DevicePrefixNamesInitializer, we preserve backwards
compatibility until all gateways are upgraded accordingly. Thus old
enum names (except of "Invalid") are still valid for parse().

The backwards compatibility is tested inside the GWMessageTest where
the "Fitprotocol" is string is parsed. However, we always expect that
the such enum is printed as "fitp".

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>