SequentialAsyncExecutor: implement finalize of tasks
When the SequentialAsyncExecutor was requested to finish, it did
not execute the queued tasks anymore. This can have some unwanted
effects in certain situations.

Thus, implement finalization of tasks with timeout. This means that
based on the given timeout (negative: blocking, 0: non-blocking,
positive: timeout blocking) when it is requested to finish, it
finalizes the tasks in the queue.

The default behaviour is non-blocking - backwards compatible.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>