DIWrapper: register class without DIW suffix
Before, each class used for dependency injection was registered
as "<full-name>DIW". The DIW suffix should have prevent some
unintentional usage. In fact, the registered class is not *that
class* we would like to register but its wrapper generated by the
BEEEON_WRAPPER macro. The wrapper is named <full-name>DIW and thus
the naming is in fact right.

However, it seems that no wierd situation has ever occoured. The
"DIW" suffix is a bit strange, especially in error messages and
reports. We can live without it and just think that the registred
thing inside DIWrapperFactory is not the class but its DIW wrapper
implementation.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>