StopControl: new class
StopControl implements an idiom commonly used in almost every
StoppableRunnable or StoppableLoop implemention. It provides
a way how to signalize stop() and interrupt the loop.

For waiting in threads, a commonly used waiting idiom is implemented
as waitStoppable(). This allows to wait while checking the stop
requests as well to not wait forever.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>