GWUnpairResponse: new class
The GWUnpairRequest is using generic responses right now. However,
certain device managers may not support unpair request in the way
we are using it in BeeeOn. Especially, the Z-Wave technology does
not allow to unpair a selected device programatically. It needs
the sensor to be online for unpair. Thus, we cannot control which
device is being unpaired.

To support this wierd scenario, the GWUnpairResponse can be used.
It contains a list of devices that were unpaired by an unpair
request.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>