GWNewDeviceGroupRequest: new class
Represents a message sent by the gateway to the server after
discovering a device consisting of a group of subdevices in a sensoric
network. Example of the device may be VPT regulator that consists of
4 zones and a boiler.

The intention of the message is to inform the server about a group
of devices that are related to each other.

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>