AutoClose: inherit Loggable instead of faking it
The previous version of AutoClose was using logger for the type
Closable. That is often just "int" (file descriptor) which has
lead to confusing logs having "int" at the place of the log
source.

This way, we will always see AutoClose, which is not very useful
in certain situations but more useful then before.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>