FileTestFixture: remove directory recursively
If the testing directory still exists before the test starts,
we would remove it first. However, such operation would fail
if the removal is not recursive (removing also files in the
directory).

Any test that would not end with the working test directory
empty would fail if TEST_REPEAT=2 is specified, like:

 ##Failure Location unknown## : Error
 Test name: BeeeOn::FilesystemDeviceCacheTest::testPairUnpair
 setUp() failed
 - uncaught exception of type std::exception (or derived).
 - Directory not empty

Fixes: 7a378d0900ef ("FileTestFixture: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>