GWSetValueRequest: support operation mode
The set-value operation might be non-reliable for certain devices
or technologies. The operation mode helps to determine how to
behave in certain cases.

Usually, the set-value operation is just a one-shot operation.
It succeeds or immediately fails. This is how the TRY_ONCE mode
should be implemented. Some implementations might implement a
more aggresive mode of setting a value for TRY_HARDER mode.
For devices that often fail during set-value, it is possible to
use REPEAT_ON_FAIL to try again until a timeout exceeds.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>