DeviceDescription: support extended details about device
Each device can now advertise more details about itself. The new entries
are:

* name - custom name usually given by user (not product name)
* firmware - information about firmware (usually name and version)
* IP address - some devices communicates over IP
* MAC address - some devices use MAC address (e.g. Bluetooth)
* serial number - some devices can be identified by serial number

Those information might be helpful for the user to better identify certain
devices. It can also help gateway to locate or identify a device after
restart (and downloading the info). It is possible to do some basic checks
of the devices' firmware to warn about security issues.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>