GWResponse: parse status() as string with backwards compatibility
The status is currently transferred as a number which is not very friendly
for debugging and makes extending of the enum Status a bit messy. Introduce
parsing of status by string value - name of the enum value. This would make
the protocol easier to understand and extend.

Later, it would be possible to modify setStatus() to serialize the status
as string "accepted", "success" or "failed" instead of converting it to
a number.

GWResponseTest test fixure is added to test the status parsing separately
from other code.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>