GWResponse: introduce status SUCCESS_PARTIAL
When an operation is of kinf 1-to-many, it is possible that it succeeds
onyl partially. I.e. some of the _many_ request processors would fail
while others succeed.

For such situation, the SUCCESS_PARTIAL can be used. The request initiator
(usually the gateway server) would know that the operation was generally
successful but there were some minor issues.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>